requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask_SQLAlchemy
